requirements.txt
pyzmq
pytest-cov
python-coveralls
codecov
coverage
lmdb
cbor
ujson
numpy
cython